+++ jwt/src/main/java/me/uport/sdk/jwt/JWTTools.kt
@@ -7,7 +7,7 @@ import com.uport.sdk.signer.UportHDSigner
 import com.uport.sdk.signer.decodeJose
 import com.uport.sdk.signer.getJoseEncoded
 import me.uport.sdk.core.*
-import me.uport.sdk.did.DIDResolver
+import me.uport.sdk.uportdid.DIDResolver
 import me.uport.sdk.ethrdid.EthrDIDResolver
 import me.uport.sdk.jsonrpc.JsonRPC
 import me.uport.sdk.jwt.model.JwtHeader
